Remove IORef effect from path copying
This is really an implementation detail, so certainly it does not belong in the
static interface. The dynamic interface is arguable, as the general point is to
allow the interface to be interpreted dynamically at runtime.

That said, it is unlikely that anyone will want to dynamically interpret IORef
operations, thus the IORefDynamic is just noise, hence omitting it from both.