Revert "Add Monad* class so that we can use our API in Eff and IO."
This reverts commit 72005a9256d6fff4a1743e4bb86462c0056d80bc.

We revert this to keep the API simpler. It turns out, maintaining
a typeclass raises uncomfortable questions about how to implement
functions using typeclass methods (Typeclass vs. Eff).

It is easier to just implement in terms of Eff, and then create
helper modules for discharging Eff into IO as needed.